* commit 'a27f9980fde4138474fa0432d4f859188d0f6e13':
  spacing
  remove unneeded files, code clean up
  spacing
  fixes
  changed RedisIntegrationTest to RedisTester and cleanup
  fix
  code clean up, reordering, file renaming
  added stats to handler
  code clean-up
  Code clean-up
  Redis Client works and all integration tests pass This commit allows redis to connect to JedisClient. (Still need to write implementation for JedisCluster)
  created  databean and databean-key for testing
  added files
  Created new maven project for Redis (datarouter-redis)
  Initial commit

+++ datarouter-storage/src/main/java/com/hotpads/datarouter/test/node/basic/manyfield/BaseManyFieldIntegrationTests.java
@@ -51,7 +51,7 @@ public abstract class BaseManyFieldIntegrationTests{
 	}
 
 	private void resetTable(){
-		if(!isMemcached()){
+		if(!isMemcached() && !isRedis()){
 			mapNode.deleteAll(null);
 		}
 	}
@@ -76,6 +76,10 @@ public abstract class BaseManyFieldIntegrationTests{
 	}
 
 	public boolean isMemcached(){
+		return false;
+	}
+
+	public boolean isRedis(){
 		return false;
 	}
 
@@ -496,5 +500,4 @@ public abstract class BaseManyFieldIntegrationTests{
 	protected void recordKey(ManyFieldBeanKey key){
 		allKeys.add(key);
 	}
-
 }
